third commit

+++ base/views.py
@@ -5,11 +5,17 @@ from . forms import RoomForm
 from django.db.models import Q
 from django.contrib import messages
 from django.contrib.auth.models import User
-
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+        
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -17,9 +23,41 @@ def loginPage(request):
         except:
             messages.error(request,'User does not exist')
 
+        user = authenticate(request,username=username, password=password)
+        if user is not None:
+            login(request,user)
+            return redirect('home')
 
-    context = {}
-    return render(render, 'base/login_register.html',context)
+        else:
+            messages.error(request,'user or password not exists')
+
+    context = {'page':page}
+    return render(request, 'base/login_register.html',context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    page = 'register'
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'Error Occured during registering')
+
+    context = {'form':form}
+    return render(request, 'base/login_register.html',context)
+
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -41,8 +79,13 @@ def room(request,pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
+
+    if request.user != room.host:
+        return HttpResponse('Your Not Allowed Here!')
+
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid:
@@ -51,10 +94,14 @@ def createRoom(request):
     context = {'form':form}
     return render(request,'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Your Not Allowed Here!')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid:
@@ -63,9 +110,13 @@ def updateRoom(request, pk):
     context = {'form':form}
     return render(request,'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+      return HttpResponse('Your Not Allowed Here!')
+
     if request.method == 'POST':
         room.delete()    
         return redirect('home')
